refactor: 클래스 명명 수정, 불필요 dto 제거
