--- src/main/java/com/wanted/preonboarding/ticket/domain/dto/ReserveSystemDto.java
@@ -1,13 +1,0 @@
-package com.wanted.preonboarding.ticket.domain.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-@Builder
-public class ReserveSystemDto {
-//    private PerformanceInfo performanceInfo;
-//    private CreateReservationRequest reserveInfo;
-    private ResponseReserveQueryDto responseQueryDto;
-}
