+++ src/main/java/com/wanted/preonboarding/ticket/domain/dto/response/CreateReservationResponse.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.ticket.domain.dto;
+package com.wanted.preonboarding.ticket.domain.dto.response;
 
 import com.wanted.preonboarding.ticket.domain.entity.Performance;
 import com.wanted.preonboarding.ticket.domain.entity.Reservation;
@@ -11,7 +11,7 @@ import java.util.UUID;
 
 @Getter @Setter
 @SuperBuilder
-public class ResponseReserveQueryDto extends BaseResDto {
+public class CreateReservationResponse extends BaseResDto {
 
     private UUID performanceId; //공연ID
     private int round; //회차
@@ -22,8 +22,8 @@ public class ResponseReserveQueryDto extends BaseResDto {
     private String reservationName; //이름
     private String reservationPhoneNumber; //연락처
 
-    public static ResponseReserveQueryDto from(Reservation reservation) {
-        return ResponseReserveQueryDto.builder()
+    public static CreateReservationResponse from(Reservation reservation) {
+        return CreateReservationResponse.builder()
                 .performanceId(reservation.getPerformanceId())
                 .round(reservation.getRound())
                 .gate(reservation.getGate())
@@ -34,8 +34,8 @@ public class ResponseReserveQueryDto extends BaseResDto {
                 .build();
     }
 
-    public static ResponseReserveQueryDto of(Performance performance, Reservation reservation) {
-        return ResponseReserveQueryDto
+    public static CreateReservationResponse of(Performance performance, Reservation reservation) {
+        return CreateReservationResponse
                 .builder()
                 .performanceId(performance.getId())
                 .performanceName(performance.getName())
